malloc: cleanup casts in tst-calloc
Followup to c3d1dac96bdd10250aa37bb367d5ef8334a093a1. As pointed out by
Maciej W. Rozycki, the casts are obviously useless now.

(cherry picked from commit fc8f253d808ade5e97c93b363bd1932023e770ba)

+++ malloc/tst-calloc.c
@@ -106,8 +106,8 @@ null_test (void)
   calloc (0, 0);
   calloc (0, max_size);
   calloc (max_size, 0);
-  calloc (0, ~((size_t) zero_size));
-  calloc (~((size_t) zero_size), 0);
+  calloc (0, ~zero_size);
+  calloc (~zero_size, 0);
   DIAG_POP_NEEDS_COMMENT;
 }
